+++ database.go
@@ -1,17 +1,54 @@
 package main
 
 import (
+	"errors"
 	"fmt"
+	"os"
+	"path/filepath"
 	"time"
 
 	"github.com/iver-wharf/wharf-core/pkg/gormutil"
 	"gorm.io/driver/postgres"
+	"gorm.io/driver/sqlite"
 	"gorm.io/gorm"
 	"gorm.io/gorm/logger"
 	"gorm.io/gorm/schema"
 )
 
+var errUnsupportedDBDriver = errors.New("unsupported database driver")
+
+func dbDriverSupportsForeignKeyConstraints(driver DBDriver) bool {
+	switch driver {
+	case DBDriverSqlite:
+		return false
+	default:
+		return true
+	}
+}
+
 func openDatabase(config DBConfig) (*gorm.DB, error) {
+	log.Info().WithString("driver", string(config.Driver)).Message("Connecting to database.")
+
+	switch config.Driver {
+	case DBDriverPostgres:
+		return openDatabasePostgres(config)
+	case DBDriverSqlite:
+		return openDatabaseSqlite(config)
+	default:
+		return nil, errUnsupportedDBDriver
+	}
+}
+
+func openDatabaseSqlite(config DBConfig) (*gorm.DB, error) {
+	if err := os.MkdirAll(filepath.Dir(config.Path), os.ModePerm); err != nil {
+		return nil, fmt.Errorf("create directories for sqlite file: %w", err)
+	}
+
+	var gormConfig = getGormConfig(config)
+	return gorm.Open(sqlite.Open(config.Path), &gormConfig)
+}
+
+func openDatabasePostgres(config DBConfig) (*gorm.DB, error) {
 	const retryDelay = 2 * time.Second
 	const maxAttempts = 3
 	psqlInfo := fmt.Sprintf("host=%s port=%d user=%s password=%s dbname=postgres sslmode=disable",
@@ -20,13 +57,7 @@ func openDatabase(config DBConfig) (*gorm.DB, error) {
 		config.Username,
 		config.Password)
 
-	var gormConfig = gorm.Config{
-		NamingStrategy: schema.NamingStrategy{
-			SingularTable: true,
-		},
-		Logger: getLogger(config),
-	}
-
+	var gormConfig = getGormConfig(config)
 	var db *gorm.DB
 	var err error
 	for attempt := 1; attempt <= maxAttempts; attempt++ {
@@ -83,6 +114,15 @@ func openDatabase(config DBConfig) (*gorm.DB, error) {
 
 	err = sqlDb.Ping()
 	return db, err
+}
+
+func getGormConfig(config DBConfig) gorm.Config {
+	return gorm.Config{
+		NamingStrategy: schema.NamingStrategy{
+			SingularTable: true,
+		},
+		Logger: getLogger(config),
+	}
 }
 
 func getLogger(config DBConfig) logger.Interface {